Moving queries into query file
