+++ src/components/queries/Standard.js
@@ -1,0 +1,11 @@
+import gql from 'graphql-tag';
+
+export const GET_STANDARDS = gql`
+  {
+    allStandards {
+      title
+      description
+      id
+    }
+  }
+`;
